pyproject.toml
[tool.poetry]
name = "llm-chat"
version = "0.1.0"
description = "LLM Chat software"
authors = ["sivarohit"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
httpx = "^0.27.0"
requests = "^2.32.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
